[stdlib] Speculative optimization: use 64-bit decoding buffers
We could also choose to use UInt buffers, so they follow the word size, pending
benchmarking on 32-bit.

+++ stdlib/public/core/UTF16.swift
@@ -95,13 +95,13 @@ extension _Unicode.UTF16 : UnicodeEncoding {
   }
   
   public struct ForwardParser {
-    public typealias _Buffer = _UIntBuffer<UInt32, UInt16>
+    public typealias _Buffer = _UIntBuffer<UInt64, UInt16>
     public init() { _buffer = _Buffer() }
     public var _buffer: _Buffer
   }
   
   public struct ReverseParser {
-    public typealias _Buffer = _UIntBuffer<UInt32, UInt16>
+    public typealias _Buffer = _UIntBuffer<UInt64, UInt16>
     public init() { _buffer = _Buffer() }
     public var _buffer: _Buffer
   }
